Add tests for database lowlevel functions
We now have a split between raw functions, which take data and add
it directly to the database, and submission functions, which
check the contents before calling the raw function.

Added a high_level write method, which is a first step of
http://tickets.musicbrainz.org/browse/AB-58, though this method call
has to be moved back into the hlcalc code.

http://tickets.musicbrainz.org/browse/AB-89