Count stats in separate queries (AB-9)
The way we were counting unique queries was unusual. The way the
data seemed to be presented was that we would have a unique count
of lossless submissions, a unique count of lossy, and then a sum of
the two.  This doesn't make sense if the same track was submitted once
as lossless, and once as lossy. It would be counted as two unique
submissions.

However, due to an unusual count query, it turns out that the value for
unique(lossy) was being counted as unique lossy submissions which
are not lossless submissions. So rather than overreporting unique
total, we were underreporting unique lossy.