Added an option for remote evaluation of dataset
The user can now see an option for evaluating the dataset on their
own machines. A column "eval_location" was added to the dataset_eval_jobs
which tells where the dataset was or is to be evaluated. "local" is
for AB server and "remote" is for user's machine. Also since this
feature cannot be shown to the user right now we have introduced
a feature flag FEATURE_EVAL_LOCATION in the config file which if
set to true shows the option to the user and if false then hides it.