Separate the behaviour of database/api in pending jobs endpoint
We make the db method only return a list of jobs. The webserver
is concerned with formatting it correctly.
Change the API endpoint name to /jobs so that we can extend it in
the future. For now, the only valid parameters let us get remote
jobs in the pending state.
Clean up some documentation