Improve runserver command to work with reloader and debugger
The runserver command was broken by a flask upgrade in #308, causing
it to lose the app context when the reloader restarted the app.
Copy the flask.cli.run_command command to our manage.py so that it acts
like `flask run`, including the debugger and reloader behaviour.
Ensure that we also reload on changes to config.RELOAD_ON_FILES

Fixes AB-327