Return json only for dataset api endpoints
Split endpoints for the dataset editor into one to render html,
which redirects to the login page if the user is not logged in,
and a service endpoint for each, which returns a json error
message if not logged in.
Reuse `@auth_required` instead of creating another one.
Ensure that we use @login_required for endpoints which return html
Fix a bug in the delete endpoint to return an html error page if
the user isn't authorised, instead of json