Set the maximum bulk query number to 25 and add documentation
This number tries to seek a balance between a good number of items
and the time taken to respond to the query. Tests show that 1 item
takes about .3s while 25 take .8s and 50 takes over 1.5s
Note that this limit has to be balanced against the uwsgi buffer size,
which will have to be increased if we also want to increase this number.