AB-411: Fix exceptions when uuids have upper-case characters
With the changes in AB-406 we introduced a bug where GET for a single
item failed because postgresql returned mbids in lower-case and we
tried to use the upper/mixed case value that was passed as a parameter.

Additionally, the Flask-uuid extension only accepts lower-case uuids by
default, meaning that some of our API endpoints returned HTTP404 when
presented with upper-case MBIDs, even if they existed in the database.

Convert all uuids to lower-case when returning them, and document this.
Accept API queries using upper-case uuids.

I made no changes to the internal datasets api because we always use
uuids returned from the api/postgres here, which are always lower-case.