Don't assume that exception messages are only ascii
Fixes AB-420
In 88ef390e I replaced Exception.message with str(exception), but this
neglected to consider the case where the message in the exception was
UTF-8. In python 2, str(themessage) caused a UnicodeEncodeError.
We can use six.ensure_text to convert bytes to a unicode string, but we
can't use this directly on an Exception instance.
Out of all of our customs Exceptions, ValidationException is the one
that most uses user-input in its error messages, so update it to take an
explicit `error` field, which we can use ensure_text() on to ensure that
this message is always represented in error messages in unicode.