Select unprocessed lowlevel documents with id larger than last batch
In order to prevent a large join across lowlevel, lowlevel_json, and
high_level, keep track of the highest id processed, and use it to limit
the number of items when making a query.
Use a subquery so that the join only happens over two small tables
(lowlevel and highlevel), and then an outer query to actually select the
json documents.