Only validate filter_type if FEATURE_EVAL_FILTERING is True
If the feature flag is disabled, the HTML form doesn't include this
field, so we have to ignore the pre_validate() method, which checks if
the value is one of the valid choices (as there is no value submitted
with the form, it's invalid).
Explicitly perform this check if the feature flag is enabled.