Add configuration to set the max limit for datasets contents
Datasets with a large number of recordings take a lot of time time to
fetch and generate the zip file. This can cause timeouts, we cannot
increase timeouts on gateways as that can lead to consuming too much
resources. Hence, set a upper limit on the size of datasets that can
be downloaded.