Remove redundant exports_dir
The export_path parameter should contain the complete path where the
results should be stored. Everywhere we use exports_dir, we append it
to the export_path. If its a matter of user convenience, we can add it
back of the command line but merge it with export_path asap.