Simplify FindCreateDirectory to a function
Making a class to just call a util method does not seem useful. Also,
the directories are always created by the ClassifcationTaskManager and
DatasetExplorer so we can just refer to their path with os.path.join
elsewher. We already have a similar function in AB utils but that is
used on Python 2.7 which does not have exist_ok arg available. Hence,
keeping this.