Use a single column in similarity.eval_results table to store results
We shouldn't store data that should be consumed at the same time in two
different fields, in case they get out of order for some reason. By
replacing it with a jsonb field we can directly add the result of a
similarity lookup, and easily compare it.

This change will require updates to the db methods that read and update
this table, but because we have disabled it for now we'll skip the
change until we re-enable feedback submission.