Don't use authorization token as flask-login request auth
By using @login_manager.request_loader to log in a user based on an
authorization token, a leaked token could be used to access any user
management page. Instead, have a specific decorator for api endpoints
that can be called with an authorization token, and rely on flask-login
for all other website endpoints that require a valid session.