Don't include part number in tar location, update subdir locations
With a file acousticbrainz-lowlevel-json-20220616-0.tar.zstd, we want
the files inside to be in a subdirectory
  acousticbrainz-lowlevel-json-20220616/
so that when many part files are uncompressed in the same directory they
all fill in the same directory structure.

With 30m data files and a subdirectory structure of m/mb/mbid-offset.json
we end up with 30m/256 = 117000 files per directory, which could cause
some problems with ls taking too long.
Instead, save as mb/i/mbid-offset.json, resulting in ~7000 items per
directory